[FEAT]/#19-CREATE-BOOKMARKLIKE-make response pkg in bookmarklike
