+++ src/main/java/umc/project/umark/domain/bookmark/controller/BookMarkController.java
@@ -11,6 +11,8 @@ import umc.project.umark.domain.bookmark.entity.BookMark;
 import umc.project.umark.domain.bookmark.repository.BookMarkRepository;
 import umc.project.umark.domain.bookmark.service.BookMarkService;
 import umc.project.umark.domain.mapping.BookMarkLike;
+import umc.project.umark.domain.mapping.converter.BookMarkLikeConverter;
+import umc.project.umark.domain.mapping.dto.BookMarkLikeResponse;
 import umc.project.umark.global.common.ApiResponse;
 import umc.project.umark.global.exception.GlobalException;
 
@@ -28,10 +30,10 @@ public class BookMarkController {
     }
 
     @PostMapping("/likes")
-    public ApiResponse<BookMarkResponse.BookMarkLikeResponseDTO> BookMarkLike(@RequestParam Long bookMarkId, @RequestParam Long memberId) {
+    public ApiResponse<BookMarkLikeResponse.BookMarkLikeResponseDTO> BookMarkLike(@RequestParam Long bookMarkId, @RequestParam Long memberId) {
 
             BookMark bookMark = bookMarkService.likeBookMark(memberId, bookMarkId);
-            return ApiResponse.onSuccess(BookMarkConverter.toBookMarkLikeResponseDTO(bookMark));
+            return ApiResponse.onSuccess(BookMarkLikeConverter.toBookMarkLikeResponseDTO(bookMark));
 
     }
